move bfloat deck to new folder
add conversion functions to bfloat file

restore cov
