+++ src/VecSim/spaces/IP/IP_SSE3_BF16.h
@@ -5,7 +5,7 @@
  */
 
 #include "VecSim/spaces/space_includes.h"
-#include "VecSim/utils/types_decl.h"
+#include "VecSim/types/bfloat16.h"
 
 static inline void InnerProductLowHalfStep(__m128i v1, __m128i v2, __m128i zeros,
                                            __m128 &sum_prod) {
