+++ src/VecSim/spaces/L2/L2_AVX512BW_VBMI2_BF16.h
@@ -5,7 +5,7 @@
  */
 
 #include "VecSim/spaces/space_includes.h"
-#include "VecSim/utils/types_decl.h"
+#include "VecSim/types/bfloat16.h"
 
 static inline void L2SqrHalfStep(bfloat16 *&pVect1, bfloat16 *&pVect2, __m512 &sum,
                                  __mmask32 mask) {
